+++ library/Blunt.hs
@@ -4,6 +4,7 @@ module Blunt where
 
 import Paths_blunt (getDataFileName)
 
+import Control.Exception (SomeException, evaluate, handle)
 import Data.ByteString.Char8 (unpack)
 import Data.ByteString.Lazy (fromStrict)
 import Data.ByteString.Lazy.Char8 (pack)
@@ -18,21 +19,26 @@ main = run 8080 $ \ request respond -> do
     index <- getDataFileName "index.html"
     let method = requestMethod request
         path = pathInfo request
-        response = case (method, path) of
-            ("GET", []) -> responseFile
+    response <- case (method, path) of
+            ("GET", []) -> return $ responseFile
                 ok200
                 [("Content-Type", "text/html; charset=utf-8")]
                 index
                 Nothing
-            ("GET", ["pointfree"]) ->
+            ("GET", ["pointfree"]) -> do
                 let params = queryString request
                     input = case lookup "input" params of
                         Just (Just param) -> param
                         _ -> ""
-                    maybeOutput = pointfree' (unpack input)
-                    body = case maybeOutput of
+                maybeOutput <- pf (unpack input)
+                let body = case maybeOutput of
                         Just output -> pack output
                         Nothing -> fromStrict input
-                in  responseLBS ok200 [("Content-Type", "text/plain; charset=utf-8")] body
-            _ -> responseLBS notFound404 [] ""
+                return $ responseLBS ok200 [("Content-Type", "text/plain; charset=utf-8")] body
+            _ -> return $ responseLBS notFound404 [] ""
     respond response
+
+pf :: String -> IO (Maybe String)
+pf = handle handler . evaluate . pointfree' where
+    handler :: SomeException -> IO (Maybe String)
+    handler _ = return Nothing